move fetchPage interfaces into global file
