+++ src/interfaces.ts
@@ -1,5 +1,20 @@
+export interface AlbumInterface {
+  name: string
+  artist: string
+  image: string
+  tracks: [Track]
+  releaseDate: string
+  article: string
+  lastURL: string
+}
 
-
+interface Track {
+  name: string
+  duration: number
+  trackNum: number
+  artist: string
+  album: string
+}
 
 export interface FetchAlbumsDatum {
   artist: {name: string},
